wrote some comments for the profiler implementation
