+++ src/qrisp/jasp/interpreter_tools/interpreters/profiling_interpreter.py
@@ -16,12 +16,36 @@
 ********************************************************************************/
 """
 
+"""
+This file implements the tools to perform quantum resource estimation using Jasp
+infrastructure. The idea here is to transform the quantum instructions within a
+given Jaspr into "counting instructions". That means instead of performing some
+quantum gate, we increment an index in an array, which keeps track of how many
+instructions of each type have been performed.
+
+To do this, we implement the 
+
+qrisp.jasp.interpreter_tools.interpreters.profiling_interpreter.py
+
+Which handles the transformation logic of the Jaspr.
+This file implements the interfaces to evaluating the transformed Jaspr.
+
+"""
+
+
+from functools import lru_cache
+
 from qrisp.jasp.interpreter_tools.abstract_interpreter import insert_outvalues, extract_invalues, eval_jaxpr
 from qrisp.jasp.primitives import QuantumPrimitive, OperationPrimitive
 
-from jax.lax import while_loop, cond
+import jax
 import jax.numpy as jnp
 
+
+# This functions takes a "profiling dic", i.e. a dictionary of the form {str : int}
+# indicating what kinds of quantum gates can appear in a Jaspr.
+# It returns an equation evaluator, which increments a counter in an array for
+# each quantum operation.
 def make_profiling_eqn_evaluator(profiling_dic):
     
     from qrisp import Jaspr
@@ -32,6 +56,9 @@ def make_profiling_eqn_evaluator(profiling_dic):
         
         if isinstance(eqn.primitive, QuantumPrimitive):
             
+            # In the case of an OperationPrimitive, we determine the array index 
+            # to be increment via dictionary look-up and perform the increment
+            # via the Jax-given .at method.            
             if isinstance(eqn.primitive, OperationPrimitive):
                 
                 counting_array = invalues[-1]
@@ -45,33 +72,48 @@ def make_profiling_eqn_evaluator(profiling_dic):
                 
                 insert_outvalues(eqn, context_dic, counting_array)
             
+            # Since we don't need to track to which qubits a certain operation
+            # is applied, we can implement a really simple behavior for most 
+            # Qubit/QubitArray handling methods.
+            # We represent qubit arrays simply with integers (indicating their)
+            # size.
             elif eqn.primitive.name == "jasp.create_qubits":
+                # create_qubits has the signature (size, QuantumCircuit).
+                # Since we represent QubitArrays via integers, it is sufficient
+                # to simply return the input as the output for this primitive.
                 insert_outvalues(eqn, context_dic, invalues)
             
             elif eqn.primitive.name == "jasp.get_size":
+                # The QubitArray size is represented via an integer.
                 insert_outvalues(eqn, context_dic, invalues[0])
                 
             elif eqn.primitive.name == "jasp.fuse":
+                # The size of the fused qubit array is the size of the two added.
                 insert_outvalues(eqn, context_dic, invalues[0] + invalues[1])
                 
             elif eqn.primitive.name == "jasp.slice":
-                
+                # For the slice operation, we need to make sure, we don't go out
+                # of bounds.
                 start = jnp.max(jnp.array([invalues[1], 0]))
                 stop = jnp.min(jnp.array([invalues[2], invalues[0]]))
                 
                 insert_outvalues(eqn, context_dic, stop - start)
                 
             elif eqn.primitive.name == "jasp.get_qubit":
+                # Trivial behavior since we don't need qubit address information
                 insert_outvalues(eqn, context_dic, None)
                 
-            else:
-                if len(eqn.outvars) == 1:
-                    insert_outvalues(eqn, context_dic, invalues[-1])
-                else:
-                    insert_outvalues(eqn, context_dic, (len(eqn.outvars)-1)*[0] + [invalues[-1]])
+            elif eqn.primitive.name in ["jasp.delete_qubits", "jasp.reset"]:
+                # Trivial behavior: return the last argument (the counting array).
+                insert_outvalues(eqn, context_dic, invalues[-1])
+                
+            elif eqn.primitive.name == "jasp.measure":
+                # The measurement returns always 0
+                insert_outvalues(eqn, context_dic, [0, invalues[-1]])
                 
         elif eqn.primitive.name == "while":
             
+            # Reinterpreted body and cond function
             def body_fun(val):
                 body_res = eval_jaxpr(eqn.params["body_jaxpr"], 
                                        eqn_evaluator = profiling_eqn_evaluator)(*val)
@@ -82,19 +124,20 @@ def make_profiling_eqn_evaluator(profiling_dic):
                                        eqn_evaluator = profiling_eqn_evaluator)(*val)
                 return res
             
-            outvalues = while_loop(cond_fun, body_fun, tuple(invalues))
+            outvalues = jax.lax.while_loop(cond_fun, body_fun, tuple(invalues))
             
             insert_outvalues(eqn, context_dic, outvalues)
             
         elif eqn.primitive.name == "cond":
             
+            # Reinterpret both branches
             false_fun = eval_jaxpr(eqn.params["branches"][0], 
                                    eqn_evaluator = profiling_eqn_evaluator)
             
             true_fun = eval_jaxpr(eqn.params["branches"][1], 
                                    eqn_evaluator = profiling_eqn_evaluator)
             
-            outvalues = cond(invalues[0], true_fun, false_fun, *invalues[1:])
+            outvalues = jax.lax.cond(invalues[0], true_fun, false_fun, *invalues[1:])
             
             if not isinstance(outvalues, (list, tuple)):
                 outvalues = (outvalues, )
@@ -103,9 +146,21 @@ def make_profiling_eqn_evaluator(profiling_dic):
         
         elif eqn.primitive.name == "pjit" and isinstance(eqn.params["jaxpr"].jaxpr, Jaspr):
             
-            zipped_profiling_dic = tuple(profiling_dic.items())
+            # For qached functions, we want to make sure, the compiled function
+            # contains only a single implementation per qached function.
             
-            from qrisp.jasp.evaluation_tools import get_compiled_profiler
+            # Within a Jaspr, it is made sure that qached function, which is called
+            # multiple times only calls the Jaspr by reference in the pjit primitive
+            # this way no "copies" of the implementation appear, thus keeping
+            # the size of the intermediate representation limited.
+            
+            # We want to carry on this property. For this we use the lru_cache feature
+            # on the get_compiler_profiler function. This function returns a 
+            # jitted function, which will always return the same object if called
+            # with the same object. Since identical qached function calls are
+            # represented by the same jaxpr, we achieve our goal.
+            
+            zipped_profiling_dic = tuple(profiling_dic.items())
             
             profiler = get_compiled_profiler(eqn.params["jaxpr"].jaxpr, zipped_profiling_dic)
             
@@ -121,4 +176,16 @@ def make_profiling_eqn_evaluator(profiling_dic):
         
     return profiling_eqn_evaluator
 
+
+@lru_cache(int(1E5))           
+def get_compiled_profiler(jaxpr, zipped_profiling_dic):
     
+    profiling_dic = dict(zipped_profiling_dic)
+    
+    profiling_eqn_evaluator = make_profiling_eqn_evaluator(profiling_dic)
+    
+    @jax.jit
+    def profiler(*args):
+        return eval_jaxpr(jaxpr, eqn_evaluator = profiling_eqn_evaluator)(*args)
+    
+    return profiler
